requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.1
cssselect2==0.7.0
idna==3.4
lxml==4.9.3
nba_api==1.3.1
numpy==1.26.1
Pillow==10.1.0
PyMuPDF==1.23.5
PyMuPDFb==1.23.5
reportlab==4.0.6
requests==2.31.0
setuptools==68.0.0
svglib==1.5.1
tinycss2==1.2.1
tqdm==4.66.1
urllib3==2.0.7
webencodings==0.5.1
wheel==0.37.1
